requirements.txt
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-auth==1.23.0
httplib2==0.18.1
oauthlib==3.1.0
pytube==15.0.0
requests==2.24.0
youtube_dl==2020.11.1.1
eyed3==0.9.5
moviepy==1.0.3
python-dotenv==1.0.1
